+++ in-adapter-api/src/main/java/gohigher/support/auth/Login.java
@@ -1,4 +1,4 @@
-package gohigher.auth.support;
+package gohigher.support.auth;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
