+++ in-adapter-api/src/main/java/gohigher/support/auth/RefreshTokenCookieProvider.java
@@ -1,4 +1,4 @@
-package gohigher.auth.support;
+package gohigher.support.auth;
 
 import java.time.Duration;
 import java.util.Arrays;
