+++ in-adapter-api/src/main/java/gohigher/support/auth/TokenType.java
@@ -1,4 +1,4 @@
-package gohigher.auth.support;
+package gohigher.support.auth;
 
 public enum TokenType {
