+++ in-adapter-api/src/test/java/gohigher/support/auth/RefreshTokenCookieProviderTest.java
@@ -1,4 +1,4 @@
-package gohigher.auth.support;
+package gohigher.support.auth;
 
 import static org.assertj.core.api.Assertions.*;
